Modify: 메인 페이지, 커뮤니티 페이지 수정 및 개발 중
