+++ src/pages/Community/Community.jsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
+import TagSelect from "./InnerComponent/TagSelect";
 
 const TRAVEL_STYLE_TAGS = [
   "명소탐방",
@@ -71,58 +72,9 @@ const ACCOMMODATION_TAGS = [
 const Community = () => {
   const navigate = useNavigate();
 
-  const [clickedCount, setClickedCount] = useState([0, 0, 0]);
-
-  const [travelStyleTagColors, setTravelStyleTagColors] = useState(
-    Array(TRAVEL_STYLE_TAGS.length).fill("white")
-  );
-
-  const [restaurantTagColors, setRestaurantTagColors] = useState(
-    Array(RESTAURANT_TAGS.length).fill("white")
-  );
-
-  const [accommodationTagColors, setAccommodationTagColors] = useState(
-    Array(ACCOMMODATION_TAGS.length).fill("white")
-  );
-
-  const [selectedTags, setSelectedTags] = useState([]);
-
-  const [isTagSelectCompleteBtnClicked, setIsTagSelectCompleteBtnClicked] =
-    useState(0);
+  const [finalSixTags, setFinalSixTags] = useState([]);
 
   const [postData, setPostData] = useState([]);
-
-  // 태그 선택 완료 버튼 클릭 시 (각 카테고리당 2개씩 선택) 일어나는 동작 관리
-  useEffect(() => {
-    setSelectedTags([]);
-
-    travelStyleTagColors.forEach((tagColor, index) => {
-      if (tagColor !== "white") {
-        setSelectedTags((prevTags) => {
-          // 이전 상태를 기반으로 새로운 배열을 생성하여 요소를 추가합니다.
-          return [...prevTags, TRAVEL_STYLE_TAGS[index]];
-        });
-      }
-    });
-
-    restaurantTagColors.forEach((tagColor, index) => {
-      if (tagColor !== "white") {
-        setSelectedTags((prevTags) => {
-          // 이전 상태를 기반으로 새로운 배열을 생성하여 요소를 추가합니다.
-          return [...prevTags, RESTAURANT_TAGS[index]];
-        });
-      }
-    });
-
-    accommodationTagColors.forEach((tagColor, index) => {
-      if (tagColor !== "white") {
-        setSelectedTags((prevTags) => {
-          // 이전 상태를 기반으로 새로운 배열을 생성하여 요소를 추가합니다.
-          return [...prevTags, ACCOMMODATION_TAGS[index]];
-        });
-      }
-    });
-  }, [isTagSelectCompleteBtnClicked]);
 
   useEffect(() => {
     fetch("/data/PostData.json", {
@@ -133,6 +85,8 @@ const Community = () => {
         setPostData(data);
       });
   }, []);
+
+  useEffect(() => {}, [finalSixTags]);
 
   // 백엔드와 통신
   // useEffect(() => {
@@ -162,87 +116,10 @@ const Community = () => {
   // }, []);
 
   // 태그 클릭 시 색깔 변경
-  const handleTagClickOne = (index) => {
-    if (travelStyleTagColors[index] == "white") {
-      const newTagColors = [...travelStyleTagColors];
-      newTagColors[index] = "#04dfd9";
-      setTravelStyleTagColors(newTagColors);
-
-      const newClickedCount = [...clickedCount];
-      newClickedCount[0] += 1;
-      setClickedCount(newClickedCount);
-    } else {
-      const newTagColors = [...travelStyleTagColors];
-      newTagColors[index] = "white";
-      setTravelStyleTagColors(newTagColors);
-
-      const newClickedCount = [...clickedCount];
-      newClickedCount[0] -= 1;
-      setClickedCount(newClickedCount);
-    }
-  };
-
-  const handleTagClickTwo = (index) => {
-    if (restaurantTagColors[index] == "white") {
-      const newTagColors = [...restaurantTagColors];
-      newTagColors[index] = "#04dfd9";
-      setRestaurantTagColors(newTagColors);
-
-      const newClickedCount = [...clickedCount];
-      newClickedCount[1] += 1;
-      setClickedCount(newClickedCount);
-    } else {
-      const newTagColors = [...restaurantTagColors];
-      newTagColors[index] = "white";
-      setRestaurantTagColors(newTagColors);
-
-      const newClickedCount = [...clickedCount];
-      newClickedCount[1] -= 1;
-      setClickedCount(newClickedCount);
-    }
-  };
-
-  const handleTagClickThree = (index) => {
-    if (accommodationTagColors[index] == "white") {
-      const newTagColors = [...accommodationTagColors];
-      newTagColors[index] = "#04dfd9";
-      setAccommodationTagColors(newTagColors);
-
-      const newClickedCount = [...clickedCount];
-      newClickedCount[2] += 1;
-      setClickedCount(newClickedCount);
-    } else {
-      const newTagColors = [...accommodationTagColors];
-      newTagColors[index] = "white";
-      setAccommodationTagColors(newTagColors);
-
-      const newClickedCount = [...clickedCount];
-      newClickedCount[2] -= 1;
-      setClickedCount(newClickedCount);
-    }
-  };
-
-  const handleTagSelectCompleteBtn = () => {
-    console.log(clickedCount);
-    if (
-      clickedCount[0] !== 2 ||
-      clickedCount[1] !== 2 ||
-      clickedCount[2] !== 2
-    ) {
-      alert("태그를 카테고리 당 2개씩 골라주시기 바랍니다.");
-      return;
-    }
-
-    setIsTagSelectCompleteBtnClicked((prev) => prev + 1);
-
-    // 태그에 맞춰 게시글 필터링하기
-    // => 위의 useState에 구현했음
-
-    alert("태그에 해당하는 게시물을 불러모으는 중입니다.");
-    console.log(selectedTags);
-  };
 
   const postId = 1;
+
+  const token = localStorage.getItem("token");
 
   return (
     <Container>
@@ -251,66 +128,19 @@ const Community = () => {
         <img src="/images/main_page/main_banner.png" />
       </TitleImgBox>
       <PostBtnArea>
-        <PostBtn
-          onClick={() => {
-            navigate("/post");
-          }}
-        >
-          게시글 작성하기
-        </PostBtn>
+        {token ? (
+          <PostBtn
+            onClick={() => {
+              navigate("/post");
+            }}
+          >
+            게시글 작성하기
+          </PostBtn>
+        ) : (
+          <></>
+        )}
       </PostBtnArea>
-      <TagsArea>
-        <TagsGroupArea>
-          <TagTitleArea>
-            <TagTitle>여행스타일</TagTitle>
-          </TagTitleArea>
-          <TagSelectArea>
-            {TRAVEL_STYLE_TAGS.map((tag, index) => (
-              <TagItem
-                bgColor={travelStyleTagColors[index]}
-                onClick={() => handleTagClickOne(index)}
-              >
-                #{tag}
-              </TagItem>
-            ))}
-          </TagSelectArea>
-        </TagsGroupArea>
-        <TagsGroupArea>
-          <TagTitleArea>
-            <TagTitle>음식점</TagTitle>
-          </TagTitleArea>
-          <TagSelectArea>
-            {RESTAURANT_TAGS.map((tag, index) => (
-              <TagItem
-                bgColor={restaurantTagColors[index]}
-                onClick={() => handleTagClickTwo(index)}
-              >
-                #{tag}
-              </TagItem>
-            ))}
-          </TagSelectArea>
-        </TagsGroupArea>
-        <TagsGroupArea>
-          <TagTitleArea>
-            <TagTitle>숙소</TagTitle>
-          </TagTitleArea>
-          <TagSelectArea>
-            {ACCOMMODATION_TAGS.map((tag, index) => (
-              <TagItem
-                bgColor={accommodationTagColors[index]}
-                onClick={() => handleTagClickThree(index)}
-              >
-                #{tag}
-              </TagItem>
-            ))}
-          </TagSelectArea>
-        </TagsGroupArea>
-      </TagsArea>
-      <TagSelectCompleteArea>
-        <TagSelectCompleteBtn onClick={() => handleTagSelectCompleteBtn()}>
-          태그 선택 완료
-        </TagSelectCompleteBtn>
-      </TagSelectCompleteArea>
+      <TagSelect handleFinalSixTags={setFinalSixTags} />
       <PostShowBox>
         <PostShowArea>
           {/* {postData.map(
@@ -346,10 +176,9 @@ const Community = () => {
               </PostItem>
             );
           })}
-          {selectedTags}
         </PostShowArea>
       </PostShowBox>
-      <FooterMargin></FooterMargin>
+      <FooterMargin>{finalSixTags}</FooterMargin>
     </Container>
   );
 };
@@ -406,79 +235,6 @@ const PostBtn = styled.div`
   background-color: #04dfd9;
   /* font-family: "Freesentation-9Black"; */
 `;
-
-const TagsArea = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 2vh 0vh;
-`;
-
-const TagsGroupArea = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 50px;
-`;
-
-const TagTitleArea = styled.div`
-  display: flex;
-  width: 80vw;
-`;
-
-const TagTitle = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 135px;
-  height: 40px;
-  margin-bottom: 15px;
-  border-radius: 10px;
-  color: white;
-  background-color: #04dfd9;
-  font-family: "Freesentation-9Black";
-`;
-
-const TagSelectArea = styled.div`
-  display: grid;
-  grid-template-columns: repeat(auto-fill, 135px);
-  /* grid-template-rows: repeat(auto-fill, 50px); */
-  gap: 10px 15px;
-`;
-
-const TagItem = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 40px;
-  background-color: ${(props) => props.bgColor};
-  font-size: medium;
-  border: 1px solid grey;
-  border-radius: 5px;
-  font-family: "Freesentation-9Black";
-`;
-
-const TagSelectCompleteArea = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100vw;
-  height: 60px;
-  margin: 50px 0px;
-`;
-
-const TagSelectCompleteBtn = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 200px;
-  height: 60px;
-  border-radius: 15px;
-  background-color: #04dfd9;
-  color: white;
-`;
-
-//
 
 const PostShowBox = styled.div`
   display: flex;
